decouple stats
